Refactor all components into a single header, and make amplitude a free function
