+++ src/CMakeLists.txt
@@ -11,7 +11,7 @@ find_package(fmt CONFIG REQUIRED)
 add_executable(beluga
     beluga.m.cpp
     noise_maker.cpp
-    instrument.cpp
+    components.cpp
 )
 
 target_link_libraries(beluga PRIVATE
