--- src/instrument.cpp
@@ -1,9 +1,0 @@
-#include "instrument.h"
-#include "constants.h"
-#include "helpers.h"
-
-#include <fmt/format.h>
-
-namespace blga {
-
-}
